Update post feature

+++ backend/controllers/post.controller.js
@@ -59,12 +59,21 @@ module.exports.readOnePost = async (req, res) => {
 
 
 module.exports.updatePost = async (req, res, next) => {
+
+    const token = req.cookies.jwt;
+    const decoded = jwtAuth.verify(token, process.env.TOKEN_SECRET);
+    const userId = decoded.id;
+
     const { id } = req.params;
     const { body } = req;
     console.log(req.params);
+
+    const userFound = await User.findByPk(userId);
+
     await Post.findByPk(id)
         .then((post) => {
             if (!post) throw new UserError("Ce contenu n'existe pas !", 0);
+            if (post.userId !== userFound.id) return res.status(401).json('Vous ne pouvez pas modifier cette publication.')
             post.content = body.content;
             post.save()
                 .then(() => res.status(201).json(post))
@@ -74,16 +83,34 @@ module.exports.updatePost = async (req, res, next) => {
         })
 }
 
-module.exports.deletePost = (req, res, next) => {
+module.exports.deletePost = async (req, res, next) => {
+
+    const token = req.cookies.jwt;
+    const decoded = jwtAuth.verify(token, process.env.TOKEN_SECRET);
+    const userId = decoded.id;
+
     const { id } = req.params;
-    Post.destroy({ where: { id: id } })
-        .then((post) => {
-            if (post === 0) throw new RequestError("Ce contenu n'existe pas !")
-            res.status(200).json('Ce contenu a été suppirmé avec succès !')
-        })
-        .catch(err => {
-            next(err)
-        })
+
+    const user = await User.findByPk(userId);
+
+    await Post.findOne({
+        where: { id: id }
+    }).then((postUserFound) => {
+        if (!postUserFound) return res.status(404).json('Utilisateur non trouvé.');
+        if (postUserFound.userId !== user.id) return res.status(401).json('Vous ne pouvez pas supprimer cette publication.')
+
+        Post.destroy({ where: { id: id } })
+            .then((post) => {
+                if (post === 0) throw new RequestError("Ce contenu n'existe pas !")
+
+                res.status(200).json('Ce contenu a été suppirmé avec succès !')
+            })
+            .catch(err => {
+                next(err)
+            })
+    })
+
+
 
 }
